+++ src/main/java/ru/rosatom/documentflow/dto/DocStatisticDTO.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class DocStatisticDTO {
-    long count;
+    private long count;
 }
