+++ src/main/java/ru/rosatom/documentflow/services/DocumentProcessService.java
@@ -1,20 +1,15 @@
 package ru.rosatom.documentflow.services;
 
 import ru.rosatom.documentflow.models.DocProcess;
-import ru.rosatom.documentflow.models.DocProcessStatus;
-import ru.rosatom.documentflow.models.Document;
 import ru.rosatom.documentflow.models.ProcessUpdateRequest;
 
 import java.util.Collection;
-import java.util.Set;
 
 public interface DocumentProcessService {
 
     DocProcess createNewProcess(Long documentId, Long recipientId);
 
     void sendToApprove(ProcessUpdateRequest processUpdateRequest);
-
-    Set<Document> findDocumentsByProcessStatus(DocProcessStatus status);
 
     DocProcess findProcessById(Long processId);
