+++ src/main/java/ru/rosatom/documentflow/services/StatisticsService.java
@@ -1,0 +1,9 @@
+package ru.rosatom.documentflow.services;
+
+import ru.rosatom.documentflow.dto.DocStatisticDTO;
+
+public interface StatisticsService {
+    DocStatisticDTO getCount();
+
+    DocStatisticDTO getCountByStatus(String stringStatus);
+}
